Toolbar refactoring
